feat!: implement interior mutability
This commit implements interior mutability over
exterior mutability for WebwareClient. It also removes
the OpenCursor state, as it is no longer required and
is fully handled internally.

BREAKING CHANGE:

Due to this commit's changes, several functions are
now async and only take normal references to itself
as opposed to the previous mutable references.